requirements.txt
aiohttp 
aiosignal 
altair 
annotated-types 
anyio 
async-timeout 
attrs 
blinker 
Bottleneck 
Brotli 
cachetools 
certifi 
charset-normalizer 
click 
dataclasses-json 
exceptiongroup 
frozenlist 
gitdb 
GitPython 
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet 
grpcio==1.62.1
grpcio-status==1.62.1
idna 
Jinja2 
jsonpatch 
jsonpointer==2.0
jsonschema 
jsonschema-specifications 
langchain 
langchain-community 
langchain-core 
langchain-google-genai==1.0.1
langchain-text-splitters 
langsmith 
markdown-it-py 
MarkupSafe 
marshmallow 
mdurl 
multidict 
mypy-extensions 
numexpr 
numpy 
orjson 
packaging 
pandas 
pillow 
proto-plus==1.23.0
protobuf==4.25.3
pyarrow 
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic 
pydantic_core 
pydeck 
Pygments 
PyPDF2 
PySocks 
python-dateutil 
python-dotenv==1.0.1
pytz 
PyYAML 
referencing 
requests 
rich 
rpds-py 
rsa==4.9
setuptools==68.2.2
six 
smmap 
sniffio 
SQLAlchemy 
streamlit 
tenacity 
toml 
toolz 
tornado 
tqdm==4.66.2
typing-inspect 
typing_extensions 
tzdata 
urllib3 
wheel==0.41.2
yarl